+++ sf-cvs/trunk/audacity-src/src/PitchName.h
@@ -23,22 +23,23 @@
 // The offset to 57 is used to determine the register. 
 // Each register starts with C (e.g., for middle C and A440, 
 // it's register 4).
-float Freq2Pitch(float freq);
+double Freq2Pitch(double freq);
 
-// PitchIndex returns the [0,11] index for a float pitchNum, 
+// PitchIndex returns the [0,11] index for a double pitchNum, 
 // as per result from Freq2Pitch, corresponding to modulo 12 
 // of the integer part of (pitchNum + 0.5), so 0=C, 1=C#, etc.
-unsigned int PitchIndex(float pitchNum);
+unsigned int PitchIndex(double pitchNum);
 
 // PitchName takes pitchNum (as per result from 
 // Freq2Pitch) and returns a standard pitch/note name [C, C#, etc.). 
 // Sharps are the default, unless, bWantFlats is true.
-wxChar * PitchName(float pitchNum, bool bWantFlats = false);
+wxChar * PitchName(double pitchNum, bool bWantFlats = false);
 
 // PitchName_Absolute does the same thing as PitchName, but appends 
 // the register number, e.g., instead of "C" it will return "C4" 
 // if the pitchNum corresonds to middle C.
-wxChar * PitchName_Absolute(float pitchNum, bool bWantFlats = false);
+// Sharps are the default, unless, bWantFlats is true.
+wxChar * PitchName_Absolute(double pitchNum, bool bWantFlats = false);
 
 #endif	// __AUDACITY_PITCHNAME__
